Split implementation into separate files
The original code was meant to be a replacement for the code sample on the
Cloud Office API wiki, in the sense that it was meant to be dropped-in to
the codebase of someone who wanted to use it.  Now that the recommended
way to include the code is to reference the NuGet package, it no longer
makes sense to keep everything in one file.
