+++ Rackspace.CloudOffice/Rackspace.CloudOffice.csproj
@@ -46,7 +46,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ApiClient.cs" />
+    <Compile Include="ApiException.cs" />
+    <Compile Include="BodyEncoder.cs" />
+    <Compile Include="IApiClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Throttler.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
